+++ app/configurator/configurator-state/reducer.spec.tsx
@@ -606,7 +606,7 @@ describe("deriveFiltersFromFields", () => {
             "it": "",
           },
         },
-        "version": "5.0.0",
+        "version": "5.1.0",
       }
     `);
   });