+++ app/utils/chart-config/versioning.ts
@@ -1641,6 +1641,40 @@ export const chartConfigMigrations: Migration[] = [
       return newConfig;
     },
   },
+  {
+    from: "5.0.0",
+    to: "5.1.0",
+    description: `table chart {
+      links {
+        + enabled
+        + baseUrl
+        + componentId
+      }
+    }`,
+    up: (config) => {
+      const newConfig = { ...config, version: "5.1.0" };
+
+      if (newConfig.chartType === "table") {
+        newConfig.links = {
+          enabled: false,
+          baseUrl: "",
+          componentId: "",
+          targetComponentId: "",
+        };
+      }
+
+      return newConfig;
+    },
+    down: (config) => {
+      const newConfig = { ...config, version: "5.0.0" };
+
+      if (newConfig.chartType === "table") {
+        delete newConfig.links;
+      }
+
+      return newConfig;
+    },
+  },
 ];
 
 export const migrateChartConfig = makeMigrate<ChartConfig>(
@@ -2271,6 +2305,12 @@ export const configuratorStateMigrations: Migration[] = [
     toVersion: "5.0.0",
     fromChartConfigVersion: "4.6.0",
     toChartConfigVersion: "5.0.0",
+  }),
+  makeBumpChartConfigVersionMigration({
+    fromVersion: "5.0.0",
+    toVersion: "5.1.0",
+    fromChartConfigVersion: "5.0.0",
+    toChartConfigVersion: "5.1.0",
   }),
 ];